Fix test helper layout for PSR-1 lint
Move the branch-added command, scanner, and prefixer harnesses out of their test files so PHPCS' one-class-per-file rule can validate the PR. Keep the helpers test-only and leave production contracts unchanged instead of adding protected production methods or constructor changes.

Fold the logger processor capture into LoggingProbeCommand so the console verbosity test does not need an extra helper class. The assertions remain the same: verify which Monolog levels reach processors at each CLI verbosity.

Validated with the PSR-1 class declaration sniff, project PHPCS on extracted helper files, PHP 7.4 syntax lint, and the affected PHPUnit slice.
