+++ tests/Unit/Console/Commands/AbstractRenamespacerCommandTest.php
@@ -2,14 +2,9 @@
 
 namespace BrianHenryIE\Strauss\Console\Commands;
 
-use BrianHenryIE\Strauss\Composer\Extra\StraussConfig;
 use BrianHenryIE\Strauss\TestCase;
-use Mockery;
-use Monolog\Logger;
 use Symfony\Component\Console\Input\ArgvInput;
-use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\BufferedOutput;
-use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * @coversDefaultClass \BrianHenryIE\Strauss\Console\Commands\AbstractRenamespacerCommand
@@ -24,9 +19,7 @@ class AbstractRenamespacerCommandTest extends TestCase
      */
     public function test_logger_processors_only_receive_visible_levels(array $args, array $expectedLevels): void
     {
-        $processor = $this->runProbeCommand($args);
-
-        self::assertSame($expectedLevels, $processor->levels);
+        self::assertSame($expectedLevels, $this->runProbeCommand($args));
     }
 
     /**
@@ -45,11 +38,11 @@ class AbstractRenamespacerCommandTest extends TestCase
 
     /**
      * @param string[] $args
+     * @return string[]
      */
-    private function runProbeCommand(array $args = []): CountingLogProcessor
+    private function runProbeCommand(array $args = []): array
     {
-        $processor = new CountingLogProcessor();
-        $command = new LoggingProbeCommand($processor);
+        $command = new LoggingProbeCommand();
         $input = new ArgvInput(array_merge(['probe'], $args));
         $output = new BufferedOutput();
 
@@ -57,60 +50,6 @@ class AbstractRenamespacerCommandTest extends TestCase
 
         self::assertSame(0, $exitCode);
 
-        return $processor;
-    }
-}
-
-class LoggingProbeCommand extends AbstractRenamespacerCommand
-{
-    private CountingLogProcessor $processor;
-
-    public function __construct(CountingLogProcessor $processor)
-    {
-        $this->processor = $processor;
-        parent::__construct();
-    }
-
-    protected function configure()
-    {
-        $this->setName('probe');
-        parent::configure();
-    }
-
-    protected function execute(InputInterface $input, OutputInterface $output): int
-    {
-        /** @var StraussConfig&\Mockery\MockInterface $config */
-        $config = Mockery::mock(StraussConfig::class);
-        $config->shouldReceive('isDryRun')->andReturn($input->hasOption('dry-run') && $input->getOption('dry-run') !== false);
-        $this->config = $config;
-
-        parent::execute($input, $output);
-
-        if ($this->logger instanceof Logger) {
-            $this->logger->pushProcessor($this->processor);
-        }
-
-        $this->logger->debug('debug record');
-        $this->logger->info('info record');
-        $this->logger->notice('notice record');
-
-        return self::SUCCESS;
-    }
-}
-
-class CountingLogProcessor
-{
-    /** @var string[] */
-    public array $levels = [];
-
-    /**
-     * @param array{level_name:string} $record
-     * @return array{level_name:string}
-     */
-    public function __invoke(array $record): array
-    {
-        $this->levels[] = $record['level_name'];
-
-        return $record;
+        return $command->getProcessedLevels();
     }
 }
