+++ tests/Unit/PrefixerTest.php
@@ -3203,30 +3203,3 @@ EOD;
         $this->assertEqualsRN($expected, $result);
     }
 }
-
-class PrefixerParseErrorHarness extends Prefixer
-{
-    /**
-     * @param array<string,string> $functionReplacementMap
-     */
-    public function callReplaceFunctionsBatch(string $contents, array $functionReplacementMap): string
-    {
-        return $this->replaceFunctionsBatch($contents, $functionReplacementMap);
-    }
-
-    /**
-     * @param array<string,NamespaceSymbol> $namespaceSymbols
-     */
-    public function callReplaceConstFetchNamespacesByMap(array $namespaceSymbols, string $contents): string
-    {
-        return $this->replaceConstFetchNamespacesByMap($namespaceSymbols, $contents);
-    }
-
-    /**
-     * @param NamespaceSymbol[] $namespaceSymbols
-     */
-    public function callPrepareRelativeNamespaces(string $contents, array $namespaceSymbols): string
-    {
-        return $this->prepareRelativeNamespaces($contents, $namespaceSymbols);
-    }
-}
